# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "rado",
        "description": "Documentation for hackaton project",
        "contact": {
            "email": "mail@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/book_requests": {
            "get": {
                "tags": [
                    "Book Request"
                ],
                "summary": "Get list of Book Requests",
                "description": "Returns list of Book Requests",
                "operationId": "bookRequestList",
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookRequestJsonModel"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Book Request"
                ],
                "summary": "Create Book Request",
                "description": "Create new Book Request",
                "operationId": "bookRequestCreate",
                "requestBody": {
                    "description": "Json Content",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookRequestJsonRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookRequestJsonModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/book_requests/{id}": {
            "delete": {
                "tags": [
                    "Book Request"
                ],
                "summary": "Delete Book Request",
                "description": "Delete Book Request",
                "operationId": "bookRequestDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Book request",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Book Request"
                ],
                "summary": "Update Book Request",
                "description": "Update Book Request",
                "operationId": "bookRequestUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Book request",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Json Content",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookRequestJsonRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookRequestJsonModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/leave_requests": {
            "get": {
                "tags": [
                    "Leave Request"
                ],
                "summary": "Get list of Leave Requests",
                "description": "Returns list of Leave Requests",
                "operationId": "leaveRequestList",
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LeaveRequestJsonModel"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Leave Request"
                ],
                "summary": "Create Leave Request",
                "description": "Create new Leave Request",
                "operationId": "leaveRequestCreate",
                "requestBody": {
                    "description": "Json Content",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LeaveRequestJsonRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LeaveRequestJsonModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/leave_requests/{id}": {
            "delete": {
                "tags": [
                    "Leave Request"
                ],
                "summary": "Delete Leave Request",
                "description": "Delete Leave Request",
                "operationId": "leaveRequestDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Leave request",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LeaveRequestJsonModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Leave Request"
                ],
                "summary": "Update Leave Request",
                "description": "Update Leave Request",
                "operationId": "leaveRequestUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Leave request",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Json Content",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LeaveRequestJsonRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "title": "Api response",
                "description": "Common Api response",
                "properties": {
                    "message": {
                        "title": "Message",
                        "description": "Message",
                        "type": "string",
                        "example": "Specific response message"
                    }
                },
                "type": "object"
            },
            "LeaveRequestJsonRequest": {
                "title": "Leave Request validation request",
                "description": "Leave Request validation request",
                "properties": {
                    "user_id": {
                        "description": "User id",
                        "type": "integer"
                    },
                    "start_date": {
                        "description": "Start date",
                        "type": "string"
                    },
                    "end_date": {
                        "description": "End date",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LeaveRequestJsonModel": {
                "title": "Leave Request",
                "description": "Leave Request",
                "properties": {
                    "id": {
                        "description": "Id",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "User id",
                        "type": "integer"
                    },
                    "start_date": {
                        "description": "Start date",
                        "type": "string"
                    },
                    "end_date": {
                        "description": "End date",
                        "type": "string"
                    },
                    "create_at": {
                        "description": "Created at",
                        "type": "string"
                    },
                    "updated_at": {
                        "description": "Updated at",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "BookRequestJsonRequest": {
                "title": "Book Request validation request",
                "description": "Book Request validation request",
                "properties": {
                    "user_id": {
                        "description": "User id",
                        "type": "integer"
                    },
                    "room_id": {
                        "description": "Room id",
                        "type": "integer"
                    },
                    "from": {
                        "description": "Booked from date time",
                        "type": "string",
                        "example": "2020-12-20 20:20:20"
                    },
                    "to": {
                        "description": "Booked to date time",
                        "type": "string",
                        "example": "2020-12-20 20:20:20"
                    }
                },
                "type": "object"
            },
            "BookRequestJsonModel": {
                "title": "Book Request",
                "description": "Book Request",
                "properties": {
                    "id": {
                        "description": "Id",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "User id",
                        "type": "integer"
                    },
                    "room_id": {
                        "description": "Room id",
                        "type": "integer"
                    },
                    "from": {
                        "description": "Booked from date time",
                        "type": "string",
                        "example": "2020-12-20 20:20:20"
                    },
                    "to": {
                        "description": "Booked to date time",
                        "type": "string",
                        "example": "2020-12-20 20:20:20"
                    },
                    "create_at": {
                        "description": "Created at",
                        "type": "string"
                    },
                    "updated_at": {
                        "description": "Updated at",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
    "security": [
        []
    ]
}